Fix 'next' learning interval
Although the code computes a good interval for the next ask of a
problem, based on whether it is answered correctly or not, it uses the
current value, making the results of correctly answering lag behind the
actual answering.

Fixing this should both slow down items that are correct, as well as
shorten the interval quicker on items that are incorrect.